feat: update directory structure
